Changed variable types to prevent number format exception.

+++ src/main/java/design/Model/Food/Food.java
@@ -3,8 +3,8 @@ package design.Model.Food;
 public interface Food {
     public Integer getCalories();
     public String getName();
-    public Integer getFat();
-    public Integer getProtein();
-    public Integer getFiber();
-    public Integer getCarbs();
+    public Double getFat();
+    public Double getProtein();
+    public Double getFiber();
+    public Double getCarbs();
 }
